requirements.txt

streamlit
pymupdf
pdfplumber
pillow
python-docx
pandas
openpyxl
